+++ CHANGELOG.rst
@@ -14,6 +14,15 @@ Change Log
 Unreleased
 ~~~~~~~~~~
 
+0.2.2 - 2022-03-11
+~~~~~~~~~~~~~~~~~~
+
+Fixed
++++++
+
+* Check the response status code when fetching from ReadTheDocs.
+* Only fetch the project list once from ReadTheDocs, since it's a constant.
+
 [0.2.1] - 2022-03-07
 ~~~~~~~~~~~~~~~~~~~~